book: 코드 수정 없이 새로운 클래스를 추가하는것만으로 애플리케이션의 기능을 확장함 -> 추상화가 유연한 설계를 가능하게했군

+++ src/main/java/org/example/ch02/Screening.java
@@ -30,7 +30,7 @@ public class Screening {
         return this.sequence == sequence;
     }
 
-    public int getMovieFee() { // 기본 요금을 반환
+    public double getMovieFee() { // 기본 요금을 반환
         return movie.getFee();
     }
 }
